Move SelectInput to Input/Select
