+++ app/src/userpages/components/StreamPage/ConfirmCsvImportDialog/index.jsx
@@ -7,7 +7,7 @@ import type { StreamId } from '$shared/flowtype/stream-types'
 import type { CsvUploadState } from '$userpages/flowtype/states/stream-state'
 import ModalPortal from '$shared/components/ModalPortal'
 import Dialog from '$shared/components/Dialog'
-import Select from '$shared/components/SelectInput/Select'
+import Select from '$shared/components/Input/Select'
 import FormControlLabel from '$shared/components/FormControlLabel'
 import CoreText from '$shared/components/Input/CoreText'
 import FormControlErrors from '$shared/components/FormControlErrors'
